move PPO batch to ksim types file
